Issue-208: Moved recyclerview package to core module.
