+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/credits/adapter/CreditVH.kt
@@ -6,8 +6,8 @@ import android.widget.TextView
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.extensions.downloadPoster
 import com.kirchhoff.movies.core.extensions.inflate
+import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.ui.screens.core.credits.CreditsView
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class CreditVH(parent: ViewGroup) : BaseVH<CreditsView.CreditsInfo>(parent.inflate(R.layout.item_credit)) {
