+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/credits/adapter/CreditsAdapter.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.ui.screens.core.credits.adapter
 
 import android.view.ViewGroup
+import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.ui.screens.core.credits.CreditsView
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class CreditsAdapter(itemClickListener: OnItemClickListener<CreditsView.CreditsInfo>) :
     BaseRecyclerViewAdapter<BaseVH<CreditsView.CreditsInfo>, CreditsView.CreditsInfo>(itemClickListener = itemClickListener) {
