+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/tvs/adapter/TvVH.kt
@@ -6,8 +6,8 @@ import android.widget.TextView
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.extensions.downloadPoster
 import com.kirchhoff.movies.core.extensions.inflate
+import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.data.ui.main.UITv
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class TvVH(parent: ViewGroup) : BaseVH<UITv>(parent.inflate(R.layout.item_tv)) {
