+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/tvs/adapter/TvsListAdapter.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.ui.screens.core.tvs.adapter
 
 import android.view.ViewGroup
+import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.data.ui.main.UITv
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class TvsListAdapter(clickListener: OnItemClickListener<UITv>) :
     BaseRecyclerViewAdapter<BaseVH<UITv>, UITv>(itemClickListener = clickListener) {
