+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/PersonsFragment.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.ui.screens.main.persons
 
+import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.data.ui.core.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.details.DetailsActivity
 import com.kirchhoff.movies.ui.screens.main.persons.adapter.PersonsListAdapter
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
 import org.koin.android.viewmodel.ext.android.viewModel
 
 class PersonsFragment : PaginatedScreenFragment<UIPerson, UIPaginated<UIPerson>>(),
