+++ app/src/main/java/com/kirchhoff/movies/ui/screens/reviews/list/adapter/ReviewVH.kt
@@ -4,8 +4,8 @@ import android.view.ViewGroup
 import android.widget.TextView
 import com.kirchhoff.movies.R
 import com.kirchhoff.movies.core.extensions.inflate
+import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.data.ui.details.review.UIReview
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class ReviewVH(parent: ViewGroup) : BaseVH<UIReview>(parent.inflate(R.layout.item_review)) {
