+++ app/src/main/java/com/kirchhoff/movies/ui/screens/reviews/list/adapter/ReviewsListAdapter.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.ui.screens.reviews.list.adapter
 
 import android.view.ViewGroup
+import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.data.ui.details.review.UIReview
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
-import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class ReviewsListAdapter(clickListener: OnItemClickListener<UIReview>) :
     BaseRecyclerViewAdapter<BaseVH<UIReview>, UIReview>(itemClickListener = clickListener) {
